requirements.txt
spacy==2.2.4
Flask==1.0.2
pandas==0.24.2
torch-1.3.1+cpu-cp37-cp37m-linux_x86_64.whl
nltk==3.4
praw==6.5.1
gunicorn==20.0.4
